requirements.txt
flask==1.1.2
gunicorn==20.0.4
pytest==6.2.2
requests==2.25.1
pymongo==3.11.3
python-dotenv==0.15.0
pytest-cov==2.11.1
mock==4.0.3
pytest-mock==3.5.1